CustomMiningSolution: change type of nonce

+++ src/mining/mining.h
@@ -49,7 +49,7 @@ struct CustomMiningTask
 struct CustomMiningSolution
 {
     unsigned long long taskIndex; // should match the index from task
-    unsigned int nonce;         // xmrig::JobResult.nonce
+    unsigned long long nonce;   // xmrig::JobResult.nonce
     m256i result;               // xmrig::JobResult.result, 32 bytes
 };
